Improvements to equality check, pretty printing, added subtraction, added a calculator
The equality check now checks for alpha-equivalence. Basically, the naming of variables is arbitrary and shouldn't matter if they're bound.

The pretty printer got added, it's used for debugging (or sometimes for other things if you're a masochist)

Subtraction operator added. Pretty simple, its just repeated addition.

Calculator added. It just repeatedly asks for an expression and evaluates it. It fails on expressions that at any point in evaluation grow too large, probably due to memmory constraints or a stackoverflow. I'm going to choose to call this a feature and not a bug.